api/requirements.txt
amqp==5.0.2
appnope==0.1.2
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.6
awsebcli==3.19.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
blessed==1.18.0
blinker==1.4
botocore==1.19.63
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.1.1
celery==5.0.5
cement==2.8.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
configparser==5.0.1
crayons==0.4.0
crontab==0.22.9
cryptography==3.2.1
decorator==4.4.2
Django==3.0.5
django-health-check==3.16.3
django-livereload-server==0.3.2
django-rest==0.8.6
djangorestframework==3.12.2
djongo==1.3.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
fire==0.3.1
firebase-admin==4.4.0
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
future==0.16.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
lazy-object-proxy==1.4.3
line-profiler==3.1.0
livereload==2.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
oauth2client==4.1.3
oauthlib==3.1.0
paramiko==2.7.2
parso==0.8.1
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.14.0
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.7.2
pylint==2.6.0
pymongo==3.11.3
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
schedule==0.6.0
selenium==3.141.0
semantic-version==2.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.1.9
webdriver-manager==3.2.2
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
